examples/dubbing/e2e-example/backend/requirements.txt
flask==3.0.3
elevenlabs==1.2.1
python-dotenv==1.0.1
flask-cors==4.0.1